Update v-page-title directive for Vue 3
