+++ web/src/main.ts
@@ -7,6 +7,9 @@
 // Plugins
 import { registerPlugins } from '@/plugins'
 
+// Directives
+import { registerDirectives } from '@/directives'
+
 // Components
 import App from './App.vue'
 
@@ -16,5 +19,6 @@ import { createApp } from 'vue'
 const app = createApp(App)
 
 registerPlugins(app)
+registerDirectives(app)
 
 app.mount('#app')
